+++ callisto_core/delivery/view_partials.py
@@ -22,6 +22,7 @@ and should not define:
 
 '''
 import logging
+import re
 
 import ratelimit.mixins
 from nacl.exceptions import CryptoError
@@ -29,6 +30,7 @@ from nacl.exceptions import CryptoError
 from django.conf import settings
 from django.core.exceptions import PermissionDenied
 from django.http import HttpResponse
+from django.shortcuts import redirect
 from django.urls import reverse, reverse_lazy
 from django.views import generic as views
 
@@ -58,13 +60,20 @@ class _PassphrasePartial(
         return self.storage_helper(self)
 
 
-class PassphraseClearingPartial(
+class _PassphraseClearingPartial(
+    EvalDataMixin,
     _PassphrasePartial,
 ):
 
     def get(self, request, *args, **kwargs):
         self.storage.clear_passphrases()
         return super().get(request, *args, **kwargs)
+
+
+class DashboardPartial(
+    _PassphraseClearingPartial,
+):
+    EVAL_ACTION_TYPE = 'DASHBOARD'
 
 
 ###################
@@ -78,7 +87,7 @@ class ReportBasePartial(
 ):
     model = models.Report
     storage_helper = view_helpers.EncryptedReportStorageHelper
-    EVAL_ACTION_TYPE = 'VIEW'  # fallback action, used rarely (if ever)
+    EVAL_ACTION_TYPE = 'VIEW'
 
     @property
     def site_id(self):
@@ -164,12 +173,20 @@ class _ReportAccessPartial(
         else:
             return False
 
+    def _passphrase_next_url(self, request):
+        next_url = None
+        if 'next' in request.GET:
+            if re.search('^/[\W/-]*', request.GET['next']):
+                next_url = request.GET['next']
+        return next_url
+
     def dispatch(self, request, *args, **kwargs):
         logger.debug(f'{self.__class__.__name__} access check')
-        if (
-            self.access_granted or
-            self.access_form_valid
-        ):
+
+        if (self.access_granted or self.access_form_valid) and self._passphrase_next_url(
+                request):
+            return self._redirect_from_passphrase(request)
+        elif self.access_granted or self.access_form_valid:
             return super().dispatch(request, *args, **kwargs)
         else:
             return self._render_access_form()
@@ -185,13 +202,16 @@ class _ReportAccessPartial(
         context = self.get_context_data(form=self._get_access_form())
         return self.render_to_response(context)
 
+    def _redirect_from_passphrase(self, request):
+        return redirect(self._passphrase_next_url(request))
+
     def _check_report_owner(self):
         if not self.report.owner == self.request.user:
             logger.warn(self.invalid_access_user_message)
             raise PermissionDenied
 
 
-class ReportUpdatePartial(
+class _ReportUpdatePartial(
     _ReportAccessPartial,
     views.edit.UpdateView,
 ):
@@ -209,7 +229,7 @@ class ReportUpdatePartial(
 
 
 class EncryptedWizardPartial(
-    ReportUpdatePartial,
+    _ReportUpdatePartial,
     wizard_builder_partials.WizardPartial,
 ):
     steps_helper = view_helpers.ReportStepsHelper
@@ -228,8 +248,8 @@ class EncryptedWizardPartial(
 ###################
 
 
-class ReportActionPartial(
-    ReportUpdatePartial,
+class _ReportActionPartial(
+    _ReportUpdatePartial,
 ):
     success_url = reverse_lazy('dashboard')
 
@@ -247,7 +267,7 @@ class ReportActionPartial(
 
 
 class ReportDeletePartial(
-    ReportActionPartial,
+    _ReportActionPartial,
 ):
     EVAL_ACTION_TYPE = 'DELETE'
 
@@ -256,8 +276,9 @@ class ReportDeletePartial(
 
 
 class WizardPDFPartial(
-    ReportActionPartial,
+    _ReportActionPartial,
 ):
+    EVAL_ACTION_TYPE = 'ACCESS_PDF'
 
     def form_valid(self, form):
         super().form_valid(form)