+++ callisto_core/notification/api.py
@@ -1,12 +1,23 @@
+import copy
 import logging
+import os
 import typing
 
 import gnupg
+import requests
+from reportlab.lib.enums import TA_CENTER
+from reportlab.lib.styles import getSampleStyleSheet
+from reportlab.lib.units import inch
+from reportlab.platypus import Image, PageBreak, Paragraph, Spacer
 
 from django.conf import settings
-from django.core.mail import EmailMessage
+from django.contrib.auth.tokens import default_token_generator
+from django.contrib.sites.models import Site
 from django.template import Context, Template
+from django.template.loader import get_template
 from django.utils import timezone
+from django.utils.encoding import force_bytes
+from django.utils.http import urlsafe_base64_encode
 
 from callisto_core.reporting.report_delivery import (
     PDFFullReport, PDFMatchReport,
@@ -17,38 +28,41 @@ logger = logging.getLogger(__name__)
 
 class CallistoCoreNotificationApi(object):
 
-    report_filename = "report_{0}.pdf.gpg"
-    report_title = 'Report'
+    report_filename = "callisto_report_{0}.pdf.gpg"
+    report_title = 'Callisto Record'
+    logo_path = '../../assets/callisto_logo.png'
+
+    # utilities
 
     @property
-    def from_email(self):
-        return '"Reports" <reports@{0}>'.format(self.mail_domain)
+    def ALERT_LIST(self):
+        return [
+            'tech@projectcallisto.org',
+        ]
 
     @property
-    def mail_domain(_):
-        return settings.APP_URL
+    def mail_domain(self):
+        return 'mail.callistocampus.org'
 
     @property
-    def model(_):
+    def model(self):
         from callisto_core.notification.models import EmailNotification
         return EmailNotification
 
-    def get_cover_page(self, *args, **kwargs):
-        '''TODO: create pdf api, move this there'''
-        return []
+    @property
+    def site_id(self):
+        return self.context['site_id']
+
+    @property
+    def from_email(self):
+        return f'"Callisto" <noreply@{self.mail_domain}>'
+
+    @property
+    def mailgun_post_route(self):
+        return f"https://api.mailgun.net/v3/{self.mail_domain}/messages"
 
     def user_site_id(self, user):
-        '''
-        Redefine this method and change the value of 1 to the user's site id
-
-        examples:
-            # for an Account model 1 to 1 with User that has a site attribute
-            user.account.site.id
-
-            # for a setup that utilizes settings.SITE_ID
-            settings.SITE_ID
-        '''
-        return 1
+        return user.account.site_id
 
     def split_addresses(self, addresses):
         if isinstance(addresses, str):
@@ -56,7 +70,77 @@ class CallistoCoreNotificationApi(object):
         else:
             return addresses
 
+    def get_cover_page(self, report_id, recipient):
+        title = f"{self.report_title} No.: {report_id}"
+
+        styles = getSampleStyleSheet()
+        headline_style = styles["Heading1"]
+        headline_style.alignment = TA_CENTER
+        headline_style.fontSize = 48
+        subtitle_style = styles["Heading2"]
+        subtitle_style.fontSize = 24
+        subtitle_style.leading = 26
+        subtitle_style.alignment = TA_CENTER
+
+        CoverPage = []
+        logo = os.path.join(
+            settings.BASE_DIR,
+            self.logo_path,
+        )
+
+        image = Image(logo, 3 * inch, 3 * inch)
+        CoverPage.append(image)
+        CoverPage.append(Spacer(1, 18))
+        CoverPage.append(Paragraph("CONFIDENTIAL", headline_style))
+        CoverPage.append(Spacer(1, 30))
+        CoverPage.append(Spacer(1, 40))
+        CoverPage.append(Paragraph(title, subtitle_style))
+        CoverPage.append(Spacer(1, 40))
+        paragraph = Paragraph(
+            f"Intended for: {recipient}, Title IX Coordinator", subtitle_style)
+        CoverPage.append(paragraph)
+        CoverPage.append(PageBreak())
+        return CoverPage
+
     # entrypoints
+
+    def slack_notification(
+        self,
+        msg: str,  # slack message
+        channel='',  # slack channel
+        type='',  # for test assertions
+    ):
+        pass
+
+    def send_with_kwargs(self, **kwargs):
+        self.context = {**kwargs}
+        self.send()
+
+    def send_confirmation(
+        self,
+        email_type: str,
+        to_addresses: typing.List[str],
+        site_id=0,
+    ) -> None:
+        '''
+        Send a matching or submission confirmation email to the user
+
+        email_type default valid options:
+            'match_confirmation'
+            'submit_confirmation'
+
+        Called if an email confirmation is requested
+        '''
+        from_email = '"Callisto Confirmation" <confirmation@{0}>'.format(
+            self.mail_domain,
+        )
+        self.context = {
+            'notification_name': email_type,
+            'to_addresses': to_addresses,
+            'site_id': site_id,
+            'from_email': from_email,
+        }
+        self.send()
 
     def send_report_to_authority(
         self,
@@ -85,31 +169,48 @@ class CallistoCoreNotificationApi(object):
         sent_report.report.submitted_to_school = timezone.now()
         sent_report.report.save()
 
-    def send_confirmation(
-        self,
-        email_type: str,
-        to_addresses: typing.List[str],
-        site_id=0,
-    ) -> None:
-        '''
-        Send a matching or submission confirmation email to the user
+    def send_student_verification_email(self, form, *args, **kwargs):
+        email = form.cleaned_data.get('email')
+        for user in form.get_users(email):
+            self.send_with_kwargs(
+                site_id=user.account.site_id,
+                to_addresses=[email],
+                email_subject='Verify your student email',
+                email_name='student_verification_email',
+                redirect_url=form.redirect_url,
+                email_template_name=form.view.email_template_name,
+                user=form.view.request.user,
+            )
 
-        email_type default valid options:
-            'match_confirmation'
-            'submit_confirmation'
+    def send_password_reset_email(self, form, *args, **kwargs):
+        email = form.cleaned_data.get('email')
+        for user in form.get_users(email):
+            self.send_with_kwargs(
+                site_id=user.account.site_id,
+                to_addresses=[email],
+                email_subject='Reset your password',
+                email_name='password_reset_email',
+                uid=args[2]['uid'],
+                protocol=args[2]['protocol'],
+                email_template_name=args[1],
+                user=args[2]['user'],
+                token=args[2]['token'],
+            )
 
-        Called if an email confirmation is requested
-        '''
-        from_email = '"Callisto Confirmation" <confirmation@{0}>'.format(
-            self.mail_domain,
+    def send_account_activation_email(self, user, email):
+        # TODO: mirror send_password_reset_email
+        self.send_with_kwargs(
+            email_template_name='callisto_core/accounts/account_activation_email.html',
+            to_addresses=[email],
+            site_id=user.account.site_id,
+            user=user,
+            uid=urlsafe_base64_encode(
+                force_bytes(user.pk)),
+            token=default_token_generator.make_token(
+                copy.copy(user)),
+            email_subject='Keep Our Community Safe with Callisto',
+            email_name='account_activation_email',
         )
-        self.context = {
-            'notification_name': email_type,
-            'to_addresses': to_addresses,
-            'site_id': site_id,
-            'from_email': from_email,
-        }
-        self.send()
 
     def send_matching_report_to_authority(
         self,
@@ -148,23 +249,18 @@ class CallistoCoreNotificationApi(object):
             match_report(MatchReport): MatchReport for which
                 a match has been found
         '''
-        from_email = '"Callisto Matching" <notification@{0}>'.format(
-            self.mail_domain,
-        )
+        from_email = f'"Callisto Matching" <notification@{self.mail_domain}>'
         user = match_report.report.owner
-
-        self.context = {
-            'notification_name': 'match_notification',
-            'to_addresses': [match_report.report.contact_email],
-            'site_id': self.user_site_id(user),
-            'from_email': from_email,
-            'report': match_report.report,
-            'user': user,
-        }
-        self.send()
+        self.send_with_kwargs(
+            notification_name='match_notification',
+            to_addresses=[match_report.report.contact_email],
+            site_id=self.user_site_id(user),
+            from_email=from_email,
+            report=match_report.report,
+            user=user,
+        )
 
     # report attachment
-    # TODO: write to self.attachment without dict.update
 
     def notification_with_full_report(
         self,
@@ -201,6 +297,8 @@ class CallistoCoreNotificationApi(object):
 
         self._notification_with_report(report_id, report_file, public_key)
 
+    # entrypoint helpers
+
     def _notification_with_report(self, report_id, report_file, public_key):
         report_file = self._encrypt_file(report_file, public_key)
         attachment = (
@@ -221,11 +319,9 @@ class CallistoCoreNotificationApi(object):
         ).data
 
     # send cycle
-    # TODO: make self.send execute async
-    # TODO: pass context as an arguement to send? (will help with async calls)
-    # TODO: tests a spec for pre / post (so they don't get deleted)
 
     def pre_send(self):
+        self.set_protocol()
         self.set_domain()
         self.set_notification()
         self.render_body()
@@ -251,24 +347,33 @@ class CallistoCoreNotificationApi(object):
 
     # send cycle implementation
 
-    def set_domain(self):
-        from django.contrib.sites.models import Site
-        if not self.context.get('domain'):
-            site = Site.objects.get(id=self.context.get('site_id'))
-            self.context.update({'domain': site.domain})
+    def _extra_data(self):
+        '''for tests'''
+        return {}
 
-    def set_notification(self):
-        site_id = self.context.get('site_id')
-        name = self.context['notification_name']
-        notifications = self.model.objects.on_site(site_id).filter(name=name)
-        if len(notifications) != 1:
-            logger.warn(
-                f'too many results for {self.model.__name__}(site_id={site_id}, name={name})')
-        notification = notifications[0]
-        self.context.update({
-            'subject': notification.subject,
-            'body': notification.body,
-        })
+    def _mail_attachments(self):
+        files = {'files': []}
+        if self.context.get('attachment'):
+            file_name = self.context['attachment'][0]
+            file_data = self.context['attachment'][1]
+            files['files'].append(
+                ('attachment', (file_name, file_data)),
+            )
+        return files
+
+    def set_protocol(self):
+        if not self.context.get('protocol'):
+            protocol = 'http' if settings.DEBUG else 'https'  # TODO: not this
+            self.context.update({'protocol': protocol})
+
+    def set_domain(self):
+        if not self.context.get('domain'):
+            domain = Site.objects.get(id=self.context.get('site_id')).domain
+            if settings.DEBUG:
+                domain_start = domain.split('.')[0]
+                domain_end = domain.split('.')[1] + '.' + domain.split('.')[2]
+                domain = f'{domain_start}-staging.{domain_end}'
+            self.context.update({'domain': domain})
 
     def render_body(self):
         body_template = Template(self.context['body'])
@@ -276,20 +381,66 @@ class CallistoCoreNotificationApi(object):
         body_rendered = body_template.render(body_context)
         self.context.update({'body': body_rendered})
 
+    def set_notification(self):
+        if self.context.get('email_template_name'):
+            body = get_template(
+                self.context['email_template_name']).template.source
+            self.context.update({
+                'notification_name': self.context['email_template_name'],
+                'subject': self.context['email_subject'],
+                'body': body,
+            })
+        else:
+            site_id = self.context.get('site_id')
+            name = self.context['notification_name']
+            notifications = self.model.objects.on_site(
+                site_id).filter(name=name)
+            if len(notifications) != 1:
+                logger.warn(
+                    f'too many results for {self.model.__name__}(site_id={site_id}, name={name})')
+            notification = notifications[0]
+            self.context.update({
+                'subject': notification.subject,
+                'body': notification.body,
+            })
+
     def send_email(self):
-        email = EmailMessage(
-            subject=self.context['subject'],
-            body=self.context['body'],
-            from_email=self.context.get('from_email', self.from_email),
-            to=self.split_addresses(self.context['to_addresses']),
-        )
-        if self.context.get('attachment'):
-            email.attach(*self.context.get('attachment'))
-        email.send(fail_silently=False)
-        self.context.update({'EmailMessage': email})
+        request_params = {
+            'auth': ("api", settings.MAILGUN_API_KEY),
+            'data': {
+                "from": self.context.get('from_email', self.from_email),
+                "to": self.context['to_addresses'],
+                "subject": self.context['subject'],
+                "html": self.context['body'],
+                **self._extra_data(),
+            },
+            **self._mail_attachments(),
+        }
+        response = requests.post(self.mailgun_post_route, **request_params)
+        self.context.update({
+            'response': getattr(response, 'context', response),
+            'response_status': response.status_code,
+            'response_content': response.content,
+        })
 
     def log_action(self):
         logger.info('notification.send(subject={}, name={})'.format(
             self.context['subject'],
             self.context['notification_name'],
         ))
+
+        if self.context.get('attachment'):
+            self.context.update({
+                'attachment': (
+                    self.context['attachment'][0],
+                    'FILEDATA',
+                )
+            })
+
+        if self.context.get('body'):
+            self.context.update({
+                'body': self.context['body'][:80]
+            })
+
+        if not self.context.get('response_status') == 200:
+            logger.error(f'status_code!=200, context: {self.context}')