+++ callisto_core/reporting/forms.py
@@ -2,82 +2,16 @@ import json
 import logging
 
 from django import forms
-from django.contrib.auth.forms import PasswordResetForm
-from django.urls import reverse
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode
 
 from callisto_core.delivery import (
     fields as delivery_fields, forms as delivery_forms,
     models as delivery_models,
 )
-from callisto_core.utils.api import NotificationApi, TenantApi
 from callisto_core.utils.forms import NoRequiredLabelMixin
 
-from . import fields, report_delivery, tokens, validators
+from . import fields, report_delivery, validators
 
 logger = logging.getLogger(__name__)
-
-
-class SendVerificationEmailForm(
-    NoRequiredLabelMixin,
-    PasswordResetForm,
-):
-    # only used for testing currently
-
-    def __init__(self, *args, **kwargs):
-        if kwargs.get('instance'):
-            kwargs.pop('instance')
-        self.view = kwargs.pop('view')
-        super().__init__(*args, **kwargs)
-        self.fields['email'].label = "Your school email"
-        self.fields['email'].widget.attrs.update(self.email_field_placeholder)
-
-    def clean_email(self):
-        email = self.cleaned_data.get('email')
-        validators.validate_school_email(email, request=self.view.request)
-        return email
-
-    def send_mail(self):
-        pass  # TODO
-
-    def save(self, *args, **kwargs):
-        self.send_mail()
-
-    def get_users(self, email):
-        return[self.view.request.user]
-
-    def email_field_placeholder(self):
-        school_email_domain = TenantApi.site_settings(
-            'SCHOOL_EMAIL_DOMAIN',
-            request=self.view.request,
-        )
-        return {'placeholder': ', '.join(
-            ['myname@' + x for x in school_email_domain.split(',')],
-        )},
-
-
-class ReportingVerificationEmailForm(
-    SendVerificationEmailForm,
-):
-    token_generator = tokens.StudentVerificationTokenGenerator()
-
-    def send_mail(self):
-        self.redirect_url = self._get_confirmation_url()
-        NotificationApi.send_student_verification_email(self)
-
-    def _get_confirmation_url(self):
-        uidb64 = urlsafe_base64_encode(
-            force_bytes(self.view.request.user.pk)).decode("utf-8")
-        token = self.token_generator.make_token(self.view.request.user)
-        return reverse(
-            self.view.request.resolver_match.view_name,
-            kwargs={
-                'uuid': self.view.report.uuid,
-                'uidb64': uidb64,
-                'token': token,
-            },
-        )
 
 
 class PrepForm(
@@ -90,11 +24,11 @@ class PrepForm(
         required=False,
     )
     contact_email = forms.CharField(
-        label="Email Address",
+        label="Email Address*",
         required=True,
     )
     contact_phone = forms.CharField(
-        label="Phone Number",
+        label="Phone Number*",
         required=True,
     )
     contact_notes = forms.ChoiceField(
@@ -141,7 +75,7 @@ class MatchingBaseForm(
     perp_name = forms.CharField(
         label="Perpetrator's Name",
         required=False,
-        widget=forms.TextInput(attrs={'placeholder': 'ex. John Doe'}),
+        widget=forms.TextInput(),
     )
 
     def __init__(self, *args, matching_validators=None, **kwargs):