+++ callisto_core/reporting/views.py
@@ -22,7 +22,7 @@ class ReportingSchoolEmailFormView(
     view_partials.SchoolEmailFormPartial,
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
     email_template_name = 'callisto_core/reporting/student_verification_email.html'
 
 
@@ -30,28 +30,30 @@ class ReportingSchoolEmailConfirmationView(
     view_partials.SchoolEmailConfirmationPartial,
 ):
     template_name = 'callisto_core/delivery/form.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
 
 
 class ReportingPrepView(
     view_partials.PrepPartial
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
 
 
 class ReportingMatchingView(
     view_partials.OptionalMatchingPartial
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
+    admin_email_template_name = 'callisto_core/accounts/match_confirmation_callisto_team.html'
 
 
 class ReportingConfirmationView(
     view_partials.ConfirmationPartial
 ):
     template_name = 'callisto_core/reporting/submission_confirm.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
+    admin_email_template_name = 'callisto_core/accounts/submit_confirmation_callisto_team.html'
 
 
 #################
@@ -63,33 +65,34 @@ class MatchingSchoolEmailFormView(
     view_partials.SchoolEmailFormPartial,
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
     email_template_name = 'callisto_core/reporting/student_verification_email.html'
 
 
 class MatchingSchoolEmailConfirmationView(
     view_partials.SchoolEmailConfirmationPartial,
 ):
-    template_name = 'callisto_core/delivery/form.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    template_name = 'callisto_core/delivery/form_redirect.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
 
 
 class MatchingPrepView(
     view_partials.PrepPartial
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
 
 
 class MatchingEnterView(
     view_partials.RequiredMatchingPartial
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'
+    admin_email_template_name = 'callisto_core/accounts/match_confirmation_callisto_team.html'
 
 
 class MatchingWithdrawView(
     view_partials.MatchingWithdrawPartial,
 ):
     template_name = 'callisto_core/reporting/submission.html'
-    access_template_name = 'callisto_core/delivery/form.html'
+    access_template_name = 'callisto_core/delivery/form_redirect.html'