+++ docs/HISTORY.md
@@ -1,5 +1,71 @@
 # History / Changelog
 
+## 0.20.14 (2018-01-24)
+
+* added missing '*' for required fields
+* fixed spelling error on matching form
+    * https://github.com/project-callisto/callisto-core/pull/396
+
+## 0.20.13 (2018-01-23)
+
+* made `NotificationApi.ALERT_LIST` a property
+
+## 0.20.12 (2018-01-23)
+
+* fixed error display for matching and reporting flow
+* fixed dashboard URL in navbar
+
+## 0.20.10 (2018-01-18)
+
+* use a single mail domain (in prod)
+
+## 0.20.9 (2018-01-15)
+
+* added "*" to required fields in matching form
+* removed placeholder text from matching form fields
+    * https://github.com/project-callisto/callisto-core/pull/390
+
+## 0.20.8 (2018-01-15)
+
+* added "*" to required fields in reporting forms
+    * https://github.com/project-callisto/callisto-core/pull/389
+
+## 0.20.7 (2018-01-15)
+
+* NotificationApi patches
+    * https://github.com/project-callisto/callisto-core/pull/385
+    * https://github.com/project-callisto/callisto-core/pull/387
+
+## 0.20.5 (2018-01-15)
+
+* add a `<p>` tag to record form multiple choice labels
+
+## 0.20.4 (2018-01-10)
+
+* move wizard builder's site foreign key from `Page` to `FormQuestion`
+    * https://github.com/project-callisto/callisto-core/pull/384
+
+## 0.20.2 (2018-01-05)
+
+* update notification api for campus-client's use case
+* sync account view organization with delivery and reporting
+    * https://github.com/project-callisto/callisto-core/pull/377
+
+## 0.20.1 (2018-01-05)
+
+* merge notification api from campus-client
+    * https://github.com/project-callisto/callisto-core/pull/359
+* add report and match alerting
+    * https://github.com/project-callisto/callisto-core/pull/371
+* update eval actions
+    * https://github.com/project-callisto/callisto-core/pull/372
+
+## 0.20.0 (2017-12-18)
+
+* add callisto_core/accounts application
+
+    https://github.com/project-callisto/callisto-core/pull/346
+
 ## 0.19.9 (2017-12-21)
 
 * add basic celery setup